fix memory usage

+++ src/memory.F90
@@ -2,7 +2,7 @@
 module MemoryAllocator
 !-------------------------------------------------------------------------------------------
 implicit none
-integer :: totalMemory=0
+integer(8) :: totalMemory=0
 
 contains 
 
@@ -162,7 +162,7 @@ subroutine DeallocatorI3D(array)
   return
 end subroutine 
 
-integer function GetTotalMemory() 
+integer(8) function GetTotalMemory() 
   GetTotalMemory = totalMemory
   return
 end function
